First level fixes (autofixes and minor manual fixes)

+++ app/models/assessment.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # == Schema Information
 #
 # Table name: assessments
@@ -24,12 +26,12 @@ class Assessment < ApplicationRecord
 
   # CATEGORIES constant
   CATEGORIES_TYPES = [
-    PSYCHOMETRIC = 'psychometric'.freeze,
-    ORGANISATIONAL = 'organisational'.freeze,
-    CASE_STUDY = 'case_study'.freeze,
-    THREESIXTY = 'threesixty'.freeze,
-    MINDMILL = 'mindmill'.freeze,
-    HOGAN = 'hogan'.freeze,
+    PSYCHOMETRIC = 'psychometric',
+    ORGANISATIONAL = 'organisational',
+    CASE_STUDY = 'case_study',
+    THREESIXTY = 'threesixty',
+    MINDMILL = 'mindmill',
+    HOGAN = 'hogan'
   ].freeze
   CATEGORIES = {
     psychometric: PSYCHOMETRIC,
@@ -37,7 +39,7 @@ class Assessment < ApplicationRecord
     case_study: CASE_STUDY,
     threesixty: THREESIXTY,
     mindmill: MINDMILL,
-    hogan: HOGAN,
+    hogan: HOGAN
   }.freeze
 
   # Assessments constant
@@ -48,7 +50,7 @@ class Assessment < ApplicationRecord
   }.freeze
 
   # STATUSES constant
-  STATUSES = %i(in_progress finished).freeze
+  STATUSES = %i[in_progress finished].freeze
 
   ### ASSOCIATIONS
   ##
@@ -158,8 +160,6 @@ class Assessment < ApplicationRecord
   private
 
   def delete_hogan_assessment_setting
-    if hogan_assessment_setting && !hogan?
-      hogan_assessment_setting.destroy
-    end
+    hogan_assessment_setting.destroy if hogan_assessment_setting && !hogan?
   end
 end
